API

Warning

This section is incomplete. Readers are kindly encouraged to go through the tests and source code for more information about the API usage.

SPIRV.AnnotatedModule

SPIRV.ControlTree

SPIRV.DeltaGraph

SPIRV.Image

SPIRV.Instruction

SPIRV.LayoutStrategy

SPIRV.PhysicalInstruction

SPIRV.PhysicalModule

SPIRV.Pointer

SPIRV.Pointer

SPIRV.SPIRVTarget

SPIRV.SSAValue

SPIRV.StructType

SPIRV.SupportedFeatures

SPIRV.TypeInfo

SPIRV.TypeInfo

SPIRV.TypeInfo

SPIRV.VulkanLayout

SPIRV.align

SPIRV.align

SPIRV.compute_minimal_size

SPIRV.compute_stride

SPIRV.disassemble

SPIRV.extract_bytes

SPIRV.infer

SPIRV.infer

SPIRV.spir_type

SPIRV.@load

SPIRV.AnnotatedModule — Type

Module annotated with instruction ranges for each logical SPIR-V section.

SPIRV.ControlTree — Type

Control tree.

The leaves are labeled as REGION_BLOCK regions, with the distinguishing property that they no children.

Children nodes of any given subtree are in reverse postorder according to the original control-flow graph.

SPIRV.DeltaGraph — Type

Graph whose vertices and edges remain identical after deletion of other vertices.

SPIRV.Image — Type

SPIR-V image type.

Type parameters:

Format: SPIR-V ImageFormat enumerated value.

Dim: SPIR-V Dim enumerated value.

Depth: 64-bit integer with value 0 (not a depth image), 1 (depth image) or 2 (unknown).

Arrayed: Bool indicating whether the image is a layer of an image array.

MS: Bool indicating whether the image is multisampled.

Sampled: 64-bit integer with value 0 (unknown), 1 (may be sampled) or 2 (read-write, no sampling).

SPIRV.Instruction — Type

Parsed SPIR-V instruction. It represents an instruction of the form %result_id = %opcode(%arguments...)::%type_id.

SPIRV.LayoutStrategy — Type

Layout strategy used to compute alignments, offsets and strides.

SPIRV.PhysicalInstruction — Type

SPIR-V instruction in binary format.

SPIRV.PhysicalModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.Pointer — Type

Pointer that keeps its parent around to make sure its contents stay valid.

Used to represent the mutability of data structures in the type domain.

SPIRV.Pointer — Method

Reconstruct a pointer from a memory address.

SPIRV.SPIRVTarget — Type

SPIR-V target for compilation through the Julia frontend.

SPIRV.SSAValue — Type

SSA value used in a SPIR-V context.

Differs from Core.SSAValue in that all SPIR-V constructs will use this SSAValue type to differentiate with Julia SSA values.

SPIRV.StructType — Type

SPIR-V aggregate type.

Equality is defined in terms of identity, since different aggregate types have in principle different semantics.

SPIRV.SupportedFeatures — Type

Extensions and capabilities supported by a client API.

SPIRV.TypeInfo — Type

Type information used for associating Julia-level data types with SPIR-V types and for applying offsets to pad payload bytes extracted from Julia values.

SPIRV.TypeInfo — Method

Construct a TypeInfo from the provided Julia types, generating the corresponding SPIR-V types and appropriate offset and stride information using the provided layout strategy.

This is primarily intended for testing purposes.

SPIRV.TypeInfo — Method

Extract a TypeInfo from the existing IR type mapping.

Warn

The type mapping is not copied from the IR for performance reasons. It must not be changed in the IR later on or correctness issues may appear.

SPIRV.VulkanLayout — Type

Vulkan-compatible layout strategy.

SPIRV.align — Method

Return a vector of bytes where every logical piece of data (delimited via the provided sizes) has been aligned according to the provided offsets.

SPIRV.align — Method

Return a new vector of bytes which correspond to the data array payload after each element being aligned and spaced with stride. The array stride must not be lesser than the aligned element size.

SPIRV.compute_minimal_size — Function

compute_minimal_size(T, storage_classes::Callable, is_interface::Callable, layout::LayoutStrategy)

Compute the minimal size that a type T should have, using the alignments computed by the provided layout strategy.

SPIRV.compute_stride — Function

compute_stride(T, ir::IR, layout::LayoutStrategy)

Compute the stride that an array containing T should have, using module member offset declarations and with an extra padding at the end corresponding to the alignment of T.

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to io.

SPIRV.extract_bytes — Method

Extract bytes from a Julia value, with strictly no alignment.

SPIRV.infer — Method

Run type inference on the given MethodInstance.

SPIRV.infer — Method

Run type inference on the provided target.

The internal MethodInstance of the original target gets mutated in the process.

SPIRV.spir_type — Function

Get a SPIR-V type from a Julia type, caching the mapping in the IR if one is provided.

If wrap_mutable is set to true, then a pointer with class StorageClassFunction will wrap the result.

SPIRV.@load — Macro

@load address::T
@load address[index]::T

Load a value of type T, either directly (if no index is specified) or at index - 1 elements from address. address should be a device address, i.e. a UInt64 value representing the address of a physical storage buffer.