API

SPIRV.CFG

SPIRV.DeltaGraph

SPIRV.GenericVector

SPIRV.Instruction

SPIRV.PhysicalInstruction

SPIRV.PhysicalModule

SPIRV.Pointer

SPIRV.SSAValue

SPIRV.ScalarMatrix

SPIRV.ScalarVector

SPIRV.StructType

SPIRV.SupportedFeatures

SPIRV.compact_reducible_bbs!

SPIRV.disassemble

SPIRV.infer

SPIRV.infer

SPIRV.merge_mutually_recursive!

SPIRV.rem_head_recursion!

SPIRV.CFG — Type

Control Flow Graph (CFG)

SPIRV.DeltaGraph — Type

Graph whose vertices and edges remain identical after deletion of other vertices.

SPIRV.GenericVector — Type

Statically sized mutable vector.

SPIRV.Instruction — Type

Parsed SPIR-V instruction. It represents an instruction of the form %result_id = %opcode(%arguments...)::%type_id.

SPIRV.PhysicalInstruction — Type

SPIR-V instruction in binary format.

SPIRV.PhysicalModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.Pointer — Type

Pointer that keeps its parent around to make sure it stays valid.

Used to represent the mutability of data structures in the type domain.

SPIRV.SSAValue — Type

SSA value used in a SPIR-V context.

Differs from Core.SSAValue in that all SPIR-V constructs will use this SSAValue type to differentiate with Julia SSA values.

SPIRV.ScalarMatrix — Type

Statically sized scalar-valued matrix represented as a vector of column vectors.

SPIRV.ScalarVector — Type

Statically sized vector with scalar values only.

SPIRV.StructType — Type

SPIR-V aggregate type.

Equality is defined in terms of identity, since different aggregate types have in principle different semantics.

SPIRV.SupportedFeatures — Type

Extensions and capabilities supported by a client API.

SPIRV.compact_reducible_bbs! — Method

Compact reducible basic blocks.

1 -> 2 -> 3 becomes 1

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to io.

SPIRV.infer — Method

Run type inference on the provided CFG and wrap inferred code with a new CFG.

The internal MethodInstance of the original CFG gets mutated in the process.

SPIRV.infer — Method

Run type inference on the given MethodInstance.

SPIRV.merge_mutually_recursive! — Method

Merge mutually recursive blocks into caller.

SPIRV.rem_head_recursion! — Method

Remove head-controlled recursion.