API

SPIRV.GenericInstruction

SPIRV.GenericInstruction

SPIRV.InstructionChunk

SPIRV.SPIRModule

SPIRV.disassemble

SPIRV.GenericInstruction — Type

SPIR-V instruction in human-readable format.

SPIRV.GenericInstruction — Method

GenericInstruction(inst)

Retrieve arguments to an operation inside an InstructionChunk inst, based on its operands (which form a sequence of 4-bytes words). Converts raw integer values to their final types: enumeration, string, or integer.

SPIRV.InstructionChunk — Type

SPIR-V instruction, in binary format. Essentially, an instruction is an opcode referring to an operation followed by operands, 4-bytes words that form the payload. Be aware that an argument may be represented by one or several operands, as is the case for literal values (see the specification for more details). Therefore, the number of operands is not necessarily equal to the number of arguments.

SPIRV.SPIRModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to the provided IO.