API

Warning

This section is incomplete. Readers are kindly encouraged to go through the tests and source code for more information about the API usage.

SPIRV.CFG

SPIRV.DeltaGraph

SPIRV.Image

SPIRV.Instruction

SPIRV.LayoutStrategy

SPIRV.PhysicalInstruction

SPIRV.PhysicalModule

SPIRV.Pointer

SPIRV.Pointer

SPIRV.SSAValue

SPIRV.StructType

SPIRV.SupportedFeatures

SPIRV.VulkanLayout

SPIRV.align

SPIRV.aligned_size

SPIRV.compact_reducible_bbs!

SPIRV.disassemble

SPIRV.extract_bytes

SPIRV.infer

SPIRV.infer

SPIRV.merge_mutually_recursive!

SPIRV.rem_head_recursion!

SPIRV.spir_type

SPIRV.CFG — Type

Control Flow Graph (CFG)

SPIRV.DeltaGraph — Type

Graph whose vertices and edges remain identical after deletion of other vertices.

SPIRV.Image — Type

SPIR-V image type.

Type parameters:

Dim: SPIR-V Dim enumerated value.

Depth: 64-bit integer with value 0 (not a depth image), 1 (depth image) or 2 (unknown).

Arrayed: Bool indicating whether the image is a layer of an image array.

MS: Bool indicating whether the image is multisampled.

Sampled: 64-bit integer with value 0 (unknown), 1 (may be sampled) or 2 (read-write, no sampling).

Format: SPIR-V ImageFormat enumerated value.

SPIRV.Instruction — Type

Parsed SPIR-V instruction. It represents an instruction of the form %result_id = %opcode(%arguments...)::%type_id.

SPIRV.LayoutStrategy — Type

Layout strategy used to compute alignments, offsets and strides.

SPIRV.PhysicalInstruction — Type

SPIR-V instruction in binary format.

SPIRV.PhysicalModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.Pointer — Type

Pointer that keeps its parent around to make sure its contents stay valid.

Used to represent the mutability of data structures in the type domain.

SPIRV.Pointer — Method

Reconstruct a pointer from a memory address.

SPIRV.SSAValue — Type

SSA value used in a SPIR-V context.

Differs from Core.SSAValue in that all SPIR-V constructs will use this SSAValue type to differentiate with Julia SSA values.

SPIRV.StructType — Type

SPIR-V aggregate type.

Equality is defined in terms of identity, since different aggregate types have in principle different semantics.

SPIRV.SupportedFeatures — Type

Extensions and capabilities supported by a client API.

SPIRV.VulkanLayout — Type

Vulkan-compatible layout strategy.

SPIRV.align — Method

Return a vector of bytes where every logical piece of data has been aligned according to the SPIR-V offsets specified in the IR.

SPIRV.aligned_size — Function

aligned_size(T, ir::IR, layout::LayoutStrategy)

Compute the size that a type T should have, using the alignments computed by the provided layout strategy.

Note that this may differ than the real aligned size that the IR will expect, which will use declared member offsets. If those offsets were overriden manually or obtained by other means that this layout strategy, the results will likely be different.

SPIRV.compact_reducible_bbs! — Method

Compact reducible basic blocks.

1 -> 2 -> 3 becomes 1

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to io.

SPIRV.extract_bytes — Method

Extract bytes from a Julia value, with strictly no alignment.

SPIRV.infer — Method

Run type inference on the provided CFG and wrap inferred code with a new CFG.

The internal MethodInstance of the original CFG gets mutated in the process.

SPIRV.infer — Method

Run type inference on the given MethodInstance.

SPIRV.merge_mutually_recursive! — Method

Merge mutually recursive blocks into caller.

SPIRV.rem_head_recursion! — Method

Remove head-controlled recursion.

SPIRV.spir_type — Function

Get a SPIR-V type from a Julia type, caching the mapping in the IR if one is provided.

If wrap_mutable is set to true, then a pointer with class StorageClassFunction will wrap the result.