API

SPIRV.Instruction

SPIRV.PhysicalInstruction

SPIRV.PhysicalModule

SPIRV.SSAValue

SPIRV.disassemble

SPIRV.Instruction — Type

Parsed SPIR-V instruction. It represents an instruction of the form %result_id = %opcode(%arguments...)::%type_id.

SPIRV.PhysicalInstruction — Type

SPIR-V instruction in binary format.

SPIRV.PhysicalModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.SSAValue — Type

SSA value used in a SPIR-V context.

Differs from Core.SSAValue in that all SPIR-V constructs will use this SSAValue type to differentiate with Julia SSA values.

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to io.