API

Warning

This section is incomplete. Readers are kindly encouraged to go through the tests and source code for more information about the API usage.

SPIRV.REGION_BLOCK

SPIRV.AnnotatedModule

SPIRV.ControlTree

SPIRV.DeltaGraph

SPIRV.FeatureSupport

SPIRV.IR

SPIRV.Image

SPIRV.Instruction

SPIRV.LayoutStrategy

SPIRV.Module

SPIRV.NativeLayout

SPIRV.NoPadding

SPIRV.OpCode

SPIRV.PhysicalInstruction

SPIRV.PhysicalModule

SPIRV.Pointer

SPIRV.Pointer

SPIRV.ResultID

SPIRV.SPIRVTarget

SPIRV.ShaderCompilationCache

SPIRV.ShaderLayout

SPIRV.ShaderSource

SPIRV.StructType

SPIRV.SupportedFeatures

SPIRV.TypeMetadata

SPIRV.VulkanLayout

SPIRV.disassemble

SPIRV.spir_type

SPIRV.@load

SPIRV.@store

SPIRV.AnnotatedModule — Type

Module annotated with instruction ranges for each logical SPIR-V section, suitable for read-only operations and analyses.

Any desired modifications of annotated modules should be staged and applied via a Diff.

Warn

This module should not be modified, and in particular must not have its structure affected in any way by the insertion or removal of instructions. Modifications can cause the annotations to become out of sync with the updated state of the module, yielding undefined behavior.

SPIRV.ControlTree — Type

Control tree.

The leaves are labeled as REGION_BLOCK regions, with the distinguishing property that they have no children.

Children nodes of any given subtree are in reverse postorder according to the original control-flow graph.

SPIRV.DeltaGraph — Type

Graph whose vertices and edges remain identical after deletion of other vertices.

SPIRV.FeatureSupport — Type

Abstract type expressing SPIR-V which capabilities and extensions are supported.

SPIRV.IR — Type

Intermediate representation of SPIR-V modules.

The types and constants mappings can be updated at any time without explicitly notifying of the mutation behavior (i.e. functions may not end with !), so long as additions only are performed. Such additions are currently done upon construction of GlobalsInfo. The ID counter can also be incremented without notice.

SPIRV.Image — Type

SPIR-V image type.

Type parameters:

Format: SPIR-V ImageFormat enumerated value.

Dim: SPIR-V Dim enumerated value.

Depth: 64-bit integer with value 0 (not a depth image), 1 (depth image) or 2 (unknown).

Arrayed: Bool indicating whether the image is a layer of an image array.

MS: Bool indicating whether the image is multisampled.

Sampled: 64-bit integer with value 0 (unknown), 1 (may be sampled) or 2 (read-write, no sampling).

SPIRV.Instruction — Type

Parsed SPIR-V instruction. It represents an instruction of the form %result_id = %opcode(%arguments...)::%type_id.

SPIRV.LayoutStrategy — Type

Layout strategy used to compute alignments, offsets and strides.

SPIRV.NativeLayout — Type

" Julia layout, with a special handling of mutable fields for composite types.

Mutable fields are stored as 8-byte pointers in Julia, but in SPIR-V there is no such concept of pointer. Therefore, we treat mutable fields as if they were fully inlined - that imposes a few restrictions on the behavior of such mutable objects, but is necessary to keep some form of compatibility with GPU representations. The alternative would be to completely disallow structs which contain mutable fields.

SPIRV.NoPadding — Type

Layout strategy which assumes no padding is required at all. This might be useful for maximally packing data when serialized, reducing size; or to improve performance by avoiding padding when not needed, e.g. if you already pad your structures manually upfront (only do that if you know what you're doing).

SPIRV.PhysicalInstruction — Type

SPIR-V instruction in binary format.

SPIRV.PhysicalModule — Type

SPIR-V module, as a series of headers followed by a stream of instructions. The header embeds two magic numbers, one for the module itself and one for the tool that generated it (e.g. glslang). It also contains the version of the specification applicable to the module, the maximum ID number and an optional instruction schema.

SPIRV.Pointer — Type

Pointer that keeps its parent around to make sure its contents stay valid.

Used to represent the mutability of data structures in the type domain.

SPIRV.Pointer — Method

Reconstruct a pointer from a memory address.

SPIRV.ResultID — Type

Result ID used in a SPIR-V context, following single static assignment rules for valid modules.

SPIRV.SPIRVTarget — Type

SPIR-V target for compilation through the Julia frontend.

SPIRV.ShaderCompilationCache — Type

ShaderCompilationCache()

Create a cache that may be used to associate shaders to be compiled (ShaderSpec) with compiled code (ShaderSource).

Although this cache will already significantly speed up shader creation by compiling only what is required, performance-critical applications will additionally want to use a high-level cache that maps a ShaderSource to a driver-dependent object generated by the graphics API chosen for that application. For Vulkan, that would be a Vk.ShaderModule, for example. It is then recommended to use an IdDict to perform the caching, to avoid hashing the whole source code.

Warning

This cache assumes that the same MethodInstance and ShaderInterface will result in the same compiled shader. However, if different (non-default) SPIRVInterpreters are used, compilation may yield different results. For performance reasons, and because most users will be fine with the default interpreter, the choice of interpreter is not checked; it is the user's responsibility to ensure different caches are used if different interpreters are to be used.

SPIRV.ShaderLayout — Type

Shader-compatible layout strategy, where layout information is strictly read from shader decorations.

SPIRV.ShaderSource — Type

SPIR-V shader code, with stage and entry point information.

SPIRV.StructType — Type

SPIR-V aggregate type.

Equality is defined in terms of identity, since different aggregate types have in principle different semantics.

SPIRV.SupportedFeatures — Type

Extensions and capabilities supported by a client API.

SPIRV.TypeMetadata — Type

Type metadata meant to be analyzed and modified to generate appropriate decorations.

SPIRV.VulkanLayout — Type

Vulkan-compatible layout strategy.

SPIRV.disassemble — Method

disassemble(io, spir_module)

Transform the content of spir_module into a human-readable format and prints it to io.

SPIRV.spir_type — Function

Get a SPIR-V type from a Julia type, caching the mapping in the IR if one is provided.

If wrap_mutable is set to true, then a pointer with class StorageClassFunction will wrap the result.

SPIRV.@load — Macro

@load address::T
@load address[index]::T

Load a value of type T, either directly (if no index is specified) or at index - 1 elements from address. address should be a device address, i.e. a UInt64 value representing the address of a physical storage buffer.

Note

Although @load address::T and @load address[1]::T look semantically the same, you should use whichever is appropriate given the underlying data; if you have an array pointer, use the latter, and if you have a pointer to a single element, use the former. Otherwise, executing this on the CPU will most likely crash if attempted with a T that is a mutable type, because mutable elements are stored as object pointers in arrays, requiring an extra bit of indirection that is expressed by @load address[1]::T if one wants to get the correct address.

SPIRV.@store — Macro

@store value address::T
@store value address[index]::T
@store address::T = value
@store address[index]::T = value

Store a value of type T at the given address, either directly (if no index is specified) or at an offset of index - 1 elements from address.

If value is not of type T, a conversion will be attempted.

address should be a device address, i.e. a UInt64 value representing the address of a physical storage buffer.

SPIRV.Module — Type

Logical representation of a SPIR-V module.

SPIRV.REGION_BLOCK — Constant

Sequence of blocks u ─→ [v, vs...] ─→ w

SPIRV.OpCode — Type

Enumerated value representing the type of an instruction. See https://registry.khronos.org/SPIR-V/specs/unified1/SPIRV.html#instructions3 for a list of SPIR-V opcodes (excluding extended instruction sets).